% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ml_fit}
\alias{ml_fit}
\title{The Trained Random Forest Model for m6A Prediction}
\format{
An object of class \code{randomForest}
}
\usage{
ml_fit
}
\description{
This object contains the pre-trained Random Forest model
('randomForest' class) used by the prediction functions.
}
\keyword{datasets}
