% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A Sites for Multiple Sequences}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained random forest model object}

\item{feature_df}{A data.frame containing the required features:
gc_content, RNA_type, RNA_region, exon_length, distance_to_junction,
evolutionary_conservation, DNA_5mer}

\item{positive_threshold}{Probability threshold for positive prediction (default: 0.5)}
}
\value{
A data.frame with original features plus prediction results
}
\description{
This function takes feature data and a trained model to predict m6A modification sites
for multiple sequences.
}
\examples{
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
example_data <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))
predictions <- prediction_multiple(ml_fit = rf_model, feature_df = example_data)
head(predictions)
}
