% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Encode DNA Sequence
This function encodes a DNA sequence into a suitable format for machine learning.}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character string representing the DNA sequence to be encoded.}
}
\value{
A numeric vector representing the encoded sequence.
}
\description{
Encode DNA Sequence
This function encodes a DNA sequence into a suitable format for machine learning.
}
