% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{One-hot encode DNA sequences}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA sequences. All must be the same length.}
}
\value{
A data.frame where each column represents a nucleotide position
(e.g., nt_pos1) and values are factors (A, T, C, G).
}
\description{
An internal helper function to convert a vector of DNA strings
(like 5-mers) into a one-hot encoded data frame suitable for ML models.
}
