% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{dna_coding}
\arguments{
\item{dna_strings}{DNA sequence vector}
}
\value{
The encoded data frame, with each column representing a nucleotide position.
}
\description{
Encode DNA sequence strings into position feature matrices
}
\examples{
# encoding mutiple dna sequences
sequences <- c("ATCG", "GCTA", "TTAA", "CCGG")
encoded <- dna_encoding(sequences)

# to see the results
print(encoded)
}
