% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{DNA Sequence Encoding for 5-mer}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA 5-mer sequences.}
}
\value{
A data frame with encoded nucleotide positions.
}
\description{
DNA Sequence Encoding for 5-mer
}
\examples{
if (FALSE) {
  m6APrediction:::dna_encoding(c("ATGAT", "CGTAC"))
}
}
