% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Encode DNA sequences into a matrix}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA sequences.}
}
\value{
A data frame with encoded DNA sequences.
}
\description{
This function converts DNA strings into a matrix representation,
where each nucleotide is split into columns for downstream analysis.
}
\examples{
dna_encoding(c("ATCGT", "GGTAA"))
}
