% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Nucleotide-level Encoding for DNA String Vectors}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector where each element is a pure DNA sequence (containing only "A", "T", "C", "G").
All sequences \strong{must have the same length} (the function uses the length of the first sequence as the standard).}
}
\value{
A factorized data frame with column names formatted as "nt_pos1", "nt_pos2", ..., "nt_posN" (N = length of DNA sequences).
Each column has fixed factor levels: c("A", "T", "C", "G"). Rows correspond to each DNA string in the input.
}
\description{
Splits an \strong{equal-length DNA string vector} into a factorized data frame at the nucleotide level.
Each column corresponds to a nucleotide position, and factor levels for all columns are standardized to "A", "T", "C", "G"
for subsequent feature processing in machine learning models.
}
\examples{
# Example:
example_data<-read.csv(system.file("extdata", "m6A_input_example.csv",package="m6APrediction"))
dna_5mer_vec <- example_data$DNA_5mer
dna_encoded_df <- dna_encoding(dna_5mer_vec)
head(dna_encoded_df, 3)
}
