% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{Encode DNA sequences into nucleotide position features}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector, where each element is a DNA sequence of equal length.}
}
\value{
A data frame where each column corresponds to a nucleotide position (e.g., nt_pos1, nt_pos2, ...),
and each cell contains a factor with levels "A", "T", "C", and "G".
}
\description{
This function converts a vector of DNA sequences (strings of "A", "T", "C", "G")
into a data frame representation suitable for machine learning input.
}
\examples{
dna_encoding(c("ATCGT", "TGCAT"))

}
