% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{DNA Sequence Encoding Function}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA strings, all of the same length.}
}
\value{
A data frame where each column is a factor (levels: "A", "T", "C", "G") representing a nucleotide position in the input DNA strings.
}
\description{
This function encodes DNA strings into a data frame of factor columns, where each column represents a nucleotide position.
}
\examples{
library(m6APrediction)
}
