% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{DNA Encoding Function}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A vector of DNA strings.}
}
\value{
A data frame containing the encoded sequences.
}
\description{
This function encodes DNA sequences into a numerical format.
}
\examples{
encoded_seq <- dna_encoding(c("ATCG", "GCTA"))
}
