% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{DNA序列编码函数}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{字符向量，包含DNA序列}
}
\value{
编码后的数据框
}
\description{
将DNA字符串转换为编码后的数据框
}
\examples{
dna_encoding("ATCG")
}
