% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dna_encoding}
\alias{dna_encoding}
\title{DNA Sequence Encoding}
\usage{
dna_encoding(dna_strings)
}
\arguments{
\item{dna_strings}{A character vector of DNA sequences to be encoded.}
}
\value{
A data frame where each column represents a position in the DNA sequence and each element is a factor
with levels "A", "T", "C", "G" corresponding to the nucleotides.
}
\description{
This function encodes DNA strings as a data frame where each nucleotide is represented as a factor.
Each position in the sequence is labeled as a factor with levels corresponding to the nucleotides A, T, C, and G.
}
\examples{
\dontrun{
dna_encoding(c("ATCGA", "GGGTT"))
}
}
