% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict Multiple m6A Sites
This function predicts multiple m6A sites based on input features using a trained model.}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained model (e.g., a random forest model) used for making predictions.}

\item{feature_df}{A data frame containing the features for prediction. The data frame must include:
\itemize{
\item \code{gc_content}: The GC content of the RNA sequence (numeric).
\item \code{RNA_type}: The type of RNA (character, levels: "mRNA", "lincRNA", "lncRNA", "pseudogene").
\item \code{RNA_region}: The region of the RNA sequence (character, levels: "CDS", "intron", "3'UTR", "5'UTR").
\item \code{exon_length}: The length of the exon (numeric).
\item \code{distance_to_junction}: The distance from the m6A site to the nearest splice junction (numeric).
\item \code{evolutionary_conservation}: A measure of evolutionary conservation at the m6A site (numeric).
\item \code{DNA_5mer}: The 5-mer surrounding the m6A site (character).
}}

\item{positive_threshold}{A threshold for positive predictions (numeric, default is 0.5).}
}
\value{
A vector of predictions based on the input data.
}
\description{
Predict Multiple m6A Sites
This function predicts multiple m6A sites based on input features using a trained model.
}
\examples{
model <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
input_data <- read.csv(system.file("extdata", "m6A_input_example.csv", package="m6APrediction"))
prediction_multiple(ml_fit = model, feature_df = input_data)
}
