% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Multiple Sample Prediction for m6A Sites}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained Random Forest model object.}

\item{feature_df}{A data frame containing the genomic features. Must contain columns:
"gc_content", "RNA_type", "RNA_region", "exon_length", "distance_to_junction",
"evolutionary_conservation", "DNA_5mer".}

\item{positive_threshold}{A numeric value between 0 and 1. Cutoff for "Positive" classification. Default is 0.5.}
}
\value{
The input data frame with two additional columns:
\item{predicted_m6A_prob}{The probability of the site being m6A positive.}
\item{predicted_m6A_status}{The classification ("Positive" or "Negative").}
}
\description{
Predicts the probability and status of m6A modification for multiple sites provided in a data frame.
}
\examples{
\dontrun{
# Load internal model and example data
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
input_df <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))

# Run prediction
results <- prediction_multiple(rf_model, input_df)
head(results)
}
}
