% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{predict multiple sequences for m6A sites}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{Fitted machine learning model}

\item{feature_df}{Data frame containing features}

\item{positive_threshold}{Positive class decision threshold, default value 0.5}
}
\value{
Enhanced data frame containing prediction results
}
\description{
Predicting m6A sites across multiple sequences in a dataset
}
\examples{
#load the model
ml_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))

# load the example data
example_df <- read.csv(system.file("extdata", "m6A_input_example.csv",
                                     package = "m6APrediction"))

#Perform batch forecasting
results <- prediction_multiple(ml_fit, example_df, positive_threshold = 0.6)
print(results)
}
