% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A for multiple sequences}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{Trained model (e.g., random forest).}

\item{feature_df}{Data frame with columns: \code{gc_content}, \code{RNA_type}, \code{RNA_region}, \code{exon_length}, \code{distance_to_junction}, \code{evolutionary_conservation}, \code{DNA_5mer}.}

\item{positive_threshold}{Probability threshold for "Positive" (default: 0.5).}
}
\value{
Input data frame with added \code{predicted_m6A_prob} and \code{predicted_m6A_status}.
}
\description{
Use a trained model to predict m6A status for multiple sequences.
}
\examples{

library(m6APrediction)


ml_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
feature_df <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))


prediction_multiple(ml_fit, feature_df, positive_threshold = 0.6)
}
