% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A Sites for Multiple Entries Based on Multiple Features}
\usage{
prediction_multiple(feature_df, model, positive_threshold = 0.5)
}
\arguments{
\item{feature_df}{A data frame that must contain the following columns: "gc_content", "RNA_type", "RNA_region", "exon_length", "distance_to_junction", "evolutionary_conservation", "DNA_5mer".}

\item{model}{A trained random forest model object.}

\item{positive_threshold}{A numeric value between 0 and 1, used as the cutoff to classify predictions as Positive or Negative. Default is 0.5.}
}
\value{
Returns the input data frame with two additional columns: "predicted_m6A_prob" (the probability of being positive) and "predicted_m6A_status" (the predicted status).
}
\description{
Uses a trained random forest model to predict m6A status from a data frame
containing multiple samples and features.
}
\examples{
\dontrun{
  rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
  example_df <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))
  predictions <- prediction_multiple(feature_df = example_df, model = rf_model)
  print(head(predictions))
}
}
