% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{多个样本的m6A预测}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{训练好的随机森林模型}

\item{feature_df}{包含特征的数据框}

\item{positive_threshold}{判断为正的阈值，默认0.5}
}
\value{
增加预测结果的数据框
}
\description{
多个样本的m6A预测
}
