% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Predict m6A sites for multiple sequences}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained random forest model object.}

\item{feature_df}{A data frame containing sequence features for prediction.}

\item{positive_threshold}{A numeric cutoff (default 0.5) for classification.}
}
\value{
A data frame containing prediction probabilities for m6A sites.
}
\description{
This function performs batch prediction of m6A sites using a pre-trained
random forest model included in the package.
}
\examples{
model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
input <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))
prediction_multiple(model, input)
}
