% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Multiple Samples Prediction for m6A Sites}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained machine learning model (e.g., from randomForest).}

\item{feature_df}{A data frame containing features for multiple samples. Required columns: "gc_content", "RNA_type", "RNA_region", "exon_length", "distance_to_junction", "evolutionary_conservation", "DNA_5mer".}

\item{positive_threshold}{A numeric value (default 0.5) to determine if a prediction is "Positive" (if probability exceeds this threshold).}
}
\value{
The input feature data frame with two additional columns: "predicted_m6A_prob" (predicted probability of being m6A positive) and "predicted_m6A_status" (factor with levels "Negative", "Positive").
}
\description{
This function predicts m6A status and probability for multiple samples using a trained machine learning model.
}
\examples{
 ml_fit=readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
 feature_df=read.csv(system.file("extdata", "m6A_input_example.csv",package="m6APrediction"))
 prediction_multiple(ml_fit, feature_df)
}
