% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Make multiple m6A site predictions using a trained model}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained random forest model object}

\item{feature_df}{A data frame containing features:
gc_content, RNA_type, RNA_region, exon_length,
distance_to_junction, evolutionary_conservation, DNA_5mer}

\item{positive_threshold}{Numeric (0–1) classification threshold}
}
\value{
A data frame identical to input \code{feature_df} but with two
additional columns: \code{predicted_m6A_prob} and \code{predicted_m6A_status}.
}
\description{
Predict m6A modification probabilities and statuses for multiple RNA samples.
Automatically encodes the DNA 5-mer sequences and appends the predictions.
}
\examples{
\dontrun{
model_path <- system.file("extdata", "rf_fit.rds", package="m6APrediction")
data_path  <- system.file("extdata", "m6A_input_example.csv", package="m6APrediction")
ml_fit <- readRDS(model_path)
feature_df <- read.csv(data_path, stringsAsFactors = FALSE)

# 转 factor 并指定 levels 与训练模型一致
feature_df$RNA_type <- factor(feature_df$RNA_type, levels = ml_fit$forest$xlevels$RNA_type)
feature_df$RNA_region <- factor(feature_df$RNA_region, levels = ml_fit$forest$xlevels$RNA_region)

# 调用函数
res <- prediction_multiple(ml_fit, feature_df)
head(res)
}

}
