% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Prediction for Multiple Samples}
\usage{
prediction_multiple(input_data)
}
\arguments{
\item{input_data}{A data frame of features for prediction.}
}
\value{
A data frame with predictions for each sample.
}
\description{
This function predicts outcomes for multiple samples based on input features.
}
\examples{
rf_fit <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
example_df <- read.csv(system.file("extdata", "m6A_input_example.csv", package="m6APrediction"))
predictions <- prediction_multiple(rf_fit, example_df, positive_threshold = 0.6)
}
