% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{多样本m6A位点预测}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{训练好的机器学习模型}

\item{feature_df}{包含特征的数据框}

\item{positive_threshold}{正样本阈值，默认为0.5}
}
\value{
包含预测概率和状态的数据框
}
\description{
该函数使用训练好的机器学习模型对多个样本进行m6A位点预测
}
\examples{
\dontrun{
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
example_data <- read.csv(system.file("extdata", "m6A_input_example.csv", package = "m6APrediction"))
prediction_multiple(rf_model, example_data, positive_threshold = 0.5)
}
}
