% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_multiple}
\alias{prediction_multiple}
\title{Multiple Sample Prediction for m6A Sites}
\usage{
prediction_multiple(ml_fit, feature_df, positive_threshold = 0.5)
}
\arguments{
\item{ml_fit}{A trained machine learning model (e.g., Random Forest) for predicting m6A sites.}

\item{feature_df}{A data frame containing the features for prediction. Must include columns:
"gc_content", "RNA_type", "RNA_region", "exon_length", "distance_to_junction",
"evolutionary_conservation", and "DNA_5mer".}

\item{positive_threshold}{A numeric value between 0 and 1, used to classify the predicted m6A status as "Positive"
if the probability is greater than the threshold (default is 0.5).}
}
\value{
A data frame with the same columns as \code{feature_df} plus two new columns:
"predicted_m6A_prob" (the predicted probability of m6A) and "predicted_m6A_status"
(either "Positive" or "Negative" based on the threshold).
}
\description{
This function takes a trained machine learning model and a data frame of features and predicts m6A probabilities
for multiple samples. It uses a pre-trained model to infer probabilities for each sample, assigns a status of
"Positive" or "Negative" based on a given threshold, and returns the input data frame with predicted probabilities
and status.
}
\examples{
ml_fit <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
example_df <- read.csv(system.file("extdata", "m6A_input_example.csv", package="m6APrediction"))
prediction_multiple(ml_fit, example_df, positive_threshold = 0.6)
}
