% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict a Single m6A Site
This function predicts a single m6A site based on input features using a trained model.}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained model (e.g., a random forest model) used for making predictions.}

\item{gc_content}{The GC content of the RNA sequence as a numeric value.}

\item{RNA_type}{The type of RNA (e.g., mRNA, lncRNA) as a character string.}

\item{RNA_region}{The region of the RNA sequence (e.g., coding region, UTR) as a character string.}

\item{exon_length}{The length of the exon containing the m6A site as a numeric value.}

\item{distance_to_junction}{The distance from the m6A site to the nearest splice junction as a numeric value.}

\item{evolutionary_conservation}{A measure of the evolutionary conservation at the m6A site as a numeric value.}

\item{DNA_5mer}{The 5-mer DNA sequence surrounding the m6A site as a character string.}

\item{positive_threshold}{The threshold for positive predictions, default is 0.5.}
}
\value{
A single prediction value based on the input vector.
}
\description{
Predict a Single m6A Site
This function predicts a single m6A site based on input features using a trained model.
}
\examples{
model <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
input_data <- read.csv(system.file("extdata", "m6A_input_example.csv", package="m6APrediction"))
gc_content <- input_data$gc_content
RNA_type <- input_data$RNA_type
RNA_region <- input_data$RNA_region
exon_length <- input_data$exon_length
distance_to_junction <- input_data$distance_to_junction
evolutionary_conservation <- input_data$evolutionary_conservation
DNA_5mer <- input_data$DNA_5mer
prediction_single(
  ml_fit = model,
  gc_content = gc_content,
  RNA_type = RNA_type,
  RNA_region = RNA_region,
  exon_length = exon_length,
  distance_to_junction = distance_to_junction,
  evolutionary_conservation = evolutionary_conservation,
  DNA_5mer = DNA_5mer,
  positive_threshold = 0.5
)
}
