% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Single Sample Prediction for m6A Sites}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained Random Forest model object.}

\item{gc_content}{Numeric. GC content of the site.}

\item{RNA_type}{Character. Type of RNA (e.g., "mRNA", "lincRNA").}

\item{RNA_region}{Character. Region of RNA (e.g., "CDS", "3'UTR").}

\item{exon_length}{Numeric. Length of the exon.}

\item{distance_to_junction}{Numeric. Distance to the nearest splice junction.}

\item{evolutionary_conservation}{Numeric. Conservation score.}

\item{DNA_5mer}{Character. 5-nucleotide DNA sequence (e.g., "GGACA").}

\item{positive_threshold}{Numeric. Cutoff for "Positive" classification. Default is 0.5.}
}
\value{
A named vector containing:
\item{predicted_m6A_prob}{The probability score.}
\item{predicted_m6A_status}{The classification status.}
}
\description{
Predicts the probability and status of m6A modification for a single site based on individual features.
}
\examples{
\dontrun{
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
prediction_single(rf_model, 0.5, "mRNA", "CDS", 10, 8, 0.5, "GGACA")
}
}
