% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A Modification Status for a Single Sequence}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{Trained random forest model.}

\item{gc_content}{Numeric.}

\item{RNA_type}{Character or factor (will be converted).}

\item{RNA_region}{Character or factor (will be converted).}

\item{exon_length}{Numeric.}

\item{distance_to_junction}{Numeric.}

\item{evolutionary_conservation}{Numeric.}

\item{DNA_5mer}{Character 5-mer.}

\item{positive_threshold}{Numeric threshold.}
}
\value{
Named vector with:
\itemize{
\item predicted_m6A_prob
\item predicted_m6A_status
}
}
\description{
Predicts m6A modification probability and status for a single RNA sequence.
}
\examples{
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
example_df <- read.csv(system.file("extdata", "m6A_input_example.csv",
                                   package = "m6APrediction"))

prediction_single(
  ml_fit = rf_model,
  gc_content = 0.6,
  RNA_type = "mRNA",
  RNA_region = "CDS",
  exon_length = 12,
  distance_to_junction = 5,
  evolutionary_conservation = 0.8,
  DNA_5mer = "ATCGAT"
)
}
