% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{predict single sequence for m6A sites}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{Fitted machine learning model}

\item{gc_content}{GC content}

\item{RNA_type}{RNAt ype}

\item{RNA_region}{RNA region}

\item{exon_length}{Exon length}

\item{distance_to_junction}{distance to junction}

\item{evolutionary_conservation}{evolutionary conservation}

\item{DNA_5mer}{DNA 5mer sequence}

\item{positive_threshold}{Positive class decision threshold, default value 0.5}
}
\value{
Named vectors containing prediction probabilities and states
}
\description{
predict single sequence for m6A sites
}
\examples{
#load the model
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))

 #Perform prediction on a single sequence
result <- prediction_single(
  ml_fit = rf_model,
  gc_content = 0.6,
  RNA_type = "mRNA",
  RNA_region = "CDS",
  exon_length = 12,
  distance_to_junction = 5,
  evolutionary_conservation = 0.8,
  DNA_5mer = "ATCGAT",
  positive_threshold = 0.5
)
  print(result)
}
