% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A Site for a Single Sequence}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained machine learning model.}

\item{gc_content}{Numeric. The GC content of the sequence.}

\item{RNA_type}{Character. The type of RNA (e.g., "mRNA").}

\item{RNA_region}{Character. The region of the RNA (e.g., "CDS").}

\item{exon_length}{Numeric. Length of the exon.}

\item{distance_to_junction}{Numeric. Distance to the nearest splice junction.}

\item{evolutionary_conservation}{Numeric. A score for evolutionary conservation.}

\item{DNA_5mer}{Character. The 5-mer DNA sequence.}

\item{positive_threshold}{A numeric value between 0 and 1. Probabilities above this threshold are classified as "Positive". Default is 0.5.}
}
\value{
A named vector with two elements: "predicted_m6A_prob" (the probability) and "predicted_m6A_status" ("Positive" or "Negative").
}
\description{
Predicts m6A modification status for a single sequence based on its features. It creates a one-row data frame, processes it, and returns a named vector with the prediction.
}
\examples{
\dontrun{
# Assuming 'rf_fit.rds' is in inst/extdata
model_file <- system.file("extdata", "rf_fit.rds", package="m6APrediction")
ml_model <- readRDS(model_file)

single_prediction <- prediction_single(ml_model,
                                      gc_content = 0.5,
                                      RNA_type = "mRNA",
                                      RNA_region = "CDS",
                                      exon_length = 100,
                                      distance_to_junction = 50,
                                      evolutionary_conservation = 0.8,
                                      DNA_5mer = "ATCGG")
print(single_prediction)
}
}
