% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict a Single m6A Site Based on Multiple Features}
\usage{
prediction_single(
  model,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{model}{A trained random forest model object.}

\item{gc_content}{GC content (numeric).}

\item{RNA_type}{RNA type (character, e.g., "mRNA").}

\item{RNA_region}{RNA region (character, e.g., "CDS").}

\item{exon_length}{Exon length (numeric).}

\item{distance_to_junction}{Distance to junction (numeric).}

\item{evolutionary_conservation}{Evolutionary conservation (numeric).}

\item{DNA_5mer}{A 5-character DNA sequence (e.g., "GGACA").}

\item{positive_threshold}{The cutoff for classifying as positive. Default is 0.5.}
}
\value{
A named vector containing "predicted_m6A_prob" and "predicted_m6A_status".
}
\description{
Uses a trained random forest model to predict the status of a single m6A site
based on a set of input features.
}
\examples{
\dontrun{
  rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
  single_pred <- prediction_single(
    model = rf_model,
    gc_content = 0.6,
    RNA_type = "mRNA",
    RNA_region = "CDS",
    exon_length = 12.5,
    distance_to_junction = 5,
    evolutionary_conservation = 0.8,
    DNA_5mer = "GGACU"
  )
  print(single_pred)
}
}
