% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{单个样本的m6A预测}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{训练好的模型}

\item{gc_content}{GC含量}

\item{RNA_type}{RNA类型}

\item{RNA_region}{RNA区域}

\item{exon_length}{外显子长度}

\item{distance_to_junction}{到连接点的距离}

\item{evolutionary_conservation}{保守性得分}

\item{DNA_5mer}{DNA 5碱基序列}

\item{positive_threshold}{阈值，默认0.5}
}
\value{
一个命名向量，包含预测概率和状态
}
\description{
单个样本的m6A预测
}
\examples{
\dontrun{
ml_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
prediction_single(ml_fit, gc_content = 0.6, RNA_type = "mRNA", RNA_region = "CDS",
                  exon_length = 12, distance_to_junction = 5,
                  evolutionary_conservation = 0.8, DNA_5mer = "ATCGAT",
                  positive_threshold = 0.5)
}
}
