% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Single Sample Prediction for m6A Sites}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained machine learning model (e.g., randomForest model).}

\item{gc_content}{A numeric value for the GC content.}

\item{RNA_type}{A character string specifying the RNA type (e.g., "mRNA", "lincRNA").}

\item{RNA_region}{A character string specifying the RNA region (e.g., "CDS", "intron").}

\item{exon_length}{A numeric value for the exon length.}

\item{distance_to_junction}{A numeric value for the distance to junction.}

\item{evolutionary_conservation}{A numeric value for evolutionary conservation.}

\item{DNA_5mer}{A character string representing the DNA 5-mer sequence.}

\item{positive_threshold}{A numeric value between 0 and 1. If the predicted probability is greater than this threshold, the status is "Positive"; otherwise, "Negative".}
}
\value{
A named vector with two elements: "predicted_m6A_prob" (predicted probability) and "predicted_m6A_status" (either "Positive" or "Negative").
}
\description{
This function takes individual feature values and a trained machine learning model to predict m6A probability and status for a single sample. It returns the predicted values as a named vector.
}
\examples{

ml_fit <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))

prediction_single(ml_fit, gc_content = 0.6, RNA_type = "mRNA",
                  RNA_region = "CDS", exon_length = 12,
                  distance_to_junction = 50, evolutionary_conservation = 0.8,
                  DNA_5mer = "ATGAT", positive_threshold = 0.5)
}
