% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A site for a single RNA entry}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A trained machine learning model (e.g., from \pkg{randomForest}).}

\item{gc_content}{Numeric value of GC content for the RNA sequence.}

\item{RNA_type}{Factor indicating RNA type, e.g., "mRNA", "lincRNA", "lncRNA", or "pseudogene".}

\item{RNA_region}{Factor indicating RNA region, e.g., "CDS", "intron", "3'UTR", or "5'UTR".}

\item{exon_length}{Numeric value representing the exon length.}

\item{distance_to_junction}{Numeric value for the distance to the nearest exon junction.}

\item{evolutionary_conservation}{Numeric score for evolutionary conservation.}

\item{DNA_5mer}{A character string representing the 5-mer DNA sequence.}

\item{positive_threshold}{Numeric value (default = 0.5) used as a cutoff for classification.}
}
\value{
A named vector containing:
\itemize{
\item \code{predicted_m6A_prob}: predicted probability for m6A modification.
\item \code{predicted_m6A_status}: predicted class ("Positive" or "Negative").
}
}
\description{
This function predicts m6A modification probability and classification
for a single RNA sequence input based on its features.
}
\examples{
# prediction_single(ml_fit, 0.45, "mRNA", "CDS", 1500, 120, 0.8, "ATCGA")

}
