% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Prediction for a Single Sample}
\usage{
prediction_single(single_input)
}
\arguments{
\item{single_input}{A vector of features for a single sample.}
}
\value{
The predicted outcome for the sample.
}
\description{
This function predicts the outcome for a single sample based on input features.
}
\examples{
rf_fit <- readRDS(system.file("extdata", "rf_fit.rds", package="m6APrediction"))
result <- prediction_single(c(0.5, 1, 0.3))
}
