% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{单样本m6A位点预测}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{训练好的机器学习模型}

\item{gc_content}{GC含量数值}

\item{RNA_type}{RNA类型}

\item{RNA_region}{RNA区域}

\item{exon_length}{外显子长度}

\item{distance_to_junction}{到连接点的距离}

\item{evolutionary_conservation}{进化保守性}

\item{DNA_5mer}{5-mer DNA序列}

\item{positive_threshold}{正样本阈值，默认为0.5}
}
\value{
包含预测概率和状态的命名向量
}
\description{
该函数使用训练好的机器学习模型对单个样本进行m6A位点预测
}
\examples{
\dontrun{
rf_model <- readRDS(system.file("extdata", "rf_fit.rds", package = "m6APrediction"))
prediction_single(rf_model, gc_content = 0.5, RNA_type = "mRNA",
                 RNA_region = "CDS", exon_length = 10, distance_to_junction = 8,
                 evolutionary_conservation = 0.5, DNA_5mer = "GGACA")
}
}
