% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prediction_single}
\alias{prediction_single}
\title{Predict m6A methylation for a single input site}
\usage{
prediction_single(
  ml_fit,
  gc_content,
  RNA_type,
  RNA_region,
  exon_length,
  distance_to_junction,
  evolutionary_conservation,
  DNA_5mer,
  positive_threshold = 0.5
)
}
\arguments{
\item{ml_fit}{A fitted randomForest model object (loaded from rf_fit.rds).}

\item{gc_content}{Numeric value of GC content.}

\item{RNA_type}{RNA type as a string; must be one of:
"mRNA", "lincRNA", "lncRNA", "pseudogene".}

\item{RNA_region}{RNA region as a string; must be one of:
"CDS", "intron", "3'UTR", "5'UTR".}

\item{exon_length}{Numeric exon length.}

\item{distance_to_junction}{Numeric distance to nearest exon–intron junction.}

\item{evolutionary_conservation}{Numeric conservation score.}

\item{DNA_5mer}{A 5-mer DNA sequence string.}

\item{positive_threshold}{Numeric probability threshold (default = 0.5).}
}
\value{
A named vector containing:
\itemize{
\item \code{predicted_m6A_prob}
\item \code{predicted_m6A_status}
}
}
\description{
This function performs m6A prediction for a single RNA site. It internally
constructs a one-row data frame from user-provided values and calls
\code{prediction_multiple()} to generate prediction probability and class.
}
\examples{
model <- readRDS(system.file("extdata", "rf_fit.rds",
                             package = "m6APrediction"))
pred_single_result <- prediction_single(
    ml_fit = model,
    gc_content = 0.55,
    RNA_type = "mRNA",
    RNA_region = "CDS",
    exon_length = 1500,
    distance_to_junction = 120,
    evolutionary_conservation = 0.32,
    DNA_5mer = "ATCGA"
)
print(pred_single_result)
}
